+++ DelvUI/Helpers/TextTagsHelper.cs
@@ -166,6 +166,8 @@ namespace DelvUI.Helpers
             ["[level]"] = (chara) => chara.Level > 0 ? chara.Level.ToString() : "-",
 
             ["[job]"] = (chara) => JobsHelper.JobNames.TryGetValue(chara.ClassJob.Id, out var jobName) ? jobName : "",
+            
+            ["[time-till-max-gp]"] = JobsHelper.TimeTillMaxGP,
             #endregion
         };