+++ DelvUI/Interface/Bars/BarConfig.cs
@@ -9,7 +9,7 @@ namespace DelvUI.Interface.Bars
     {
         [ColorEdit4("Background Color")]
         [Order(20)]
-        public PluginConfigColor BackgroundColor = new PluginConfigColor(new Vector4(0f / 255f, 0f / 255f, 0f / 255f, 50f / 100f));        
+        public PluginConfigColor BackgroundColor = new PluginConfigColor(new Vector4(0f / 255f, 0f / 255f, 0f / 255f, 50f / 100f));
 
         [ColorEdit4("Fill Color")]
         [Order(25)]
@@ -19,9 +19,17 @@ namespace DelvUI.Interface.Bars
         [Order(30)]
         public BarDirection FillDirection;
 
-        [Checkbox("Show Border")]
+        [Checkbox("Show Border", spacing = true)]
         [Order(35)]
         public bool DrawBorder = true;
+
+        [ColorEdit4("Border Color")]
+        [Order(36)]
+        public PluginConfigColor BorderColor = new PluginConfigColor(new Vector4(0f / 255f, 0f / 255f, 0f / 255f, 100f / 100f));
+
+        [DragInt("Border Thickness", min = 1, max = 10)]
+        [Order(37)]
+        public int BorderThickness = 1;
 
         [Checkbox("Hide When Inactive")]
         [Order(40)]