+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -50,10 +50,10 @@ namespace DelvUI.Interface.Party
             _raiseLabelHud = new LabelHud(RaiseTracker.Icon.Label);
             _invulnLabelHud = new LabelHud(InvulnTracker.Icon.Label);
 
-            _manaBarHud = new PrimaryResourceHud(_configs.ManaBar, "");
-            _castbarHud = new CastbarHud(_configs.CastBar, "");
-            _buffsListHud = new StatusEffectsListHud(_configs.Buffs, "");
-            _debuffsListHud = new StatusEffectsListHud(_configs.Debuffs, "");
+            _manaBarHud = new PrimaryResourceHud(_configs.ManaBar);
+            _castbarHud = new CastbarHud(_configs.CastBar);
+            _buffsListHud = new StatusEffectsListHud(_configs.Buffs);
+            _debuffsListHud = new StatusEffectsListHud(_configs.Debuffs);
         }
 
         public PluginConfigColor GetColor(float scale)
@@ -162,10 +162,10 @@ namespace DelvUI.Interface.Party
             {
                 bgColor = Member.InvulnStatus?.InvulnId == 811 ? InvulnTracker.WalkingDeadBackgroundColor : InvulnTracker.BackgroundColor;
             }
-            else if (_configs.HealthBar.ColorsConfig.UseDeathIndicatorBackgroundColor && Member.HP <= 0)
+            else if (_configs.HealthBar.ColorsConfig.UseDeathIndicatorBackgroundColor && Member.HP <= 0 && character != null)
             {
                 bgColor = _configs.HealthBar.RangeConfig.Enabled
-                    ? GetDistance(character, _configs.HealthBar.ColorsConfig.DeathIndicatorBackgroundColor)
+                    ? GetDistanceColor(character, _configs.HealthBar.ColorsConfig.DeathIndicatorBackgroundColor)
                     : _configs.HealthBar.ColorsConfig.DeathIndicatorBackgroundColor;
             }
 
@@ -186,7 +186,7 @@ namespace DelvUI.Interface.Party
 
             if (_configs.HealthBar.RangeConfig.Enabled)
             {
-                hpColor = GetDistance(character, hpColor);
+                hpColor = GetDistanceColor(character, hpColor);
             }
 
             DrawHelper.DrawGradientFilledRect(Position, hpFillSize, hpColor, drawList);
@@ -287,7 +287,7 @@ namespace DelvUI.Interface.Party
             }
         }
 
-        private PluginConfigColor GetDistance(Character? character, PluginConfigColor color)
+        private PluginConfigColor GetDistanceColor(Character? character, PluginConfigColor color)
         {
             byte distance = character != null ? character.YalmDistanceX : byte.MaxValue;
             float currentAlpha = color.Vector.W * 100f;
@@ -423,6 +423,7 @@ namespace DelvUI.Interface.Party
             return status switch
             {
                 PartyMemberStatus.ViewingCutscene => 61508,
+                PartyMemberStatus.Offline => 61504,
                 _ => null
             };
         }
@@ -432,6 +433,7 @@ namespace DelvUI.Interface.Party
             return status switch
             {
                 PartyMemberStatus.ViewingCutscene => "[Viewing Cutscene]",
+                PartyMemberStatus.Offline => "[Offline]",
                 _ => null
             };
         }